backendv1.0/requirements.txt
django==4.1.4
djangorestframework==3.14.0
Pillow==9.0.1